requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aioredis==2.0.1
amqp==5.1.1
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
billiard==3.6.4.0
boto3==1.26.52
botocore==1.29.52
cachetools==5.2.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.89.1
flatbuffers==23.1.4
flowdec==1.1.0
fonttools==4.38.0
future==0.18.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
oauth2client==4.1.3
gcloud==0.16.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imageio==2.24.0
importlib-metadata==6.0.0
iniconfig==2.0.0
javabridge==1.0.19
jmespath==1.0.1
jsons==1.6.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kombu==5.2.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
motor==3.1.1
mysql-connector==2.2.9
networkx==3.0
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
passlib==1.7.4
Pillow==9.4.0
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
pymongo==4.3.3
pyotp==2.8.0
pyparsing==3.0.9
pytest==7.2.1
pytest-asyncio==0.20.3
python-bioformats==4.0.7
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyWavelets==1.4.1
qrcode==7.3.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
tifffile==2022.10.10
tifftools==1.3.7
tomli==2.0.1
typeguard==2.13.3
typing-extensions==4.4.0
typish==1.9.3
urllib3==1.26.14
uuid==1.30.0
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
openslide-python
opencv-python
opencv-contrib-python
cellpose
numba
ashlar
pandas